requirements.txt
# Core
ffmpeg-python==0.2.0
yt-dlp==2025.1.15
pydantic==2.8.2
pydantic-settings
python-dotenv==1.0.1
tqdm==4.66.5
numpy==1.26.4
opencv-python-headless==4.10.0.84
pandas==2.2.2

# Audio transcription (local Whisper)
openai-whisper==20231117
torch
torchaudio

# Scenes & objects
scenedetect==0.6.3
ultralytics==8.3.34  # YOLOv8

# DB + vectors
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
sqlalchemy-pgvector

# LLM (Gemini via Google AI Studio free key)
google-generativeai==0.8.3
